Genomics_Pipeline_GH_gcp/GCP Distributed Workload/requirements.txt
pandas
numpy
pyyaml
google-cloud-storage
google-cloud-pubsub
fsspec
gcsfs